feat(suite): self-compose stablecoin yield withdraw/redeem calldata
