+++ packages/suite/src/components/earn/yield/common/YieldFlowCompleteWithdraw.tsx
@@ -1,31 +1,43 @@
 import { Translation } from '@suite/intl';
 import { type YieldFlowCompleteValue } from '@suite-common/wallet-core';
+import { Column, Row, Text } from '@trezor/components';
 
 import { YieldFlowComplete } from './YieldFlowComplete';
-import { YieldFlowTransferRow } from './YieldFlowTransferRow';
+import { YieldTokenValue } from './YieldTokenValue';
 
 type YieldFlowCompleteWithdrawProps = {
-    input: YieldFlowCompleteValue;
-    output: YieldFlowCompleteValue;
+    value: YieldFlowCompleteValue;
 };
 
-export const YieldFlowCompleteWithdraw = ({ input, output }: YieldFlowCompleteWithdrawProps) => (
+export const YieldFlowCompleteWithdraw = ({ value }: YieldFlowCompleteWithdrawProps) => (
     <YieldFlowComplete
         type="withdraw"
         heading={<Translation id="TR_EARN_YIELD_WITHDRAW_COMPLETE" />}
         description={
             <Translation
                 id="TR_EARN_YIELD_WITHDRAW_COMPLETE_DESCRIPTION"
-                values={{ displaySymbol: output.token.symbol }}
+                values={{ displaySymbol: value.token.symbol }}
             />
         }
         showFeedback
     >
-        <YieldFlowTransferRow
-            inputLabelId="TR_SENT"
-            outputLabelId="TR_RECEIVED"
-            input={input}
-            output={output}
-        />
+        <Row
+            justifyContent="space-between"
+            alignItems="center"
+            padding={{ vertical: 16, horizontal: 20 }}
+        >
+            <Column gap={8}>
+                <Text typographyStyle="body-md">
+                    <Translation id="TR_EARN_YIELD_AMOUNT_TO_WITHDRAW" />
+                </Text>
+                <YieldTokenValue
+                    token={{
+                        ...value.token,
+                        contractAddress: value.token.contractAddress ?? null,
+                    }}
+                    amount={value.amount}
+                />
+            </Column>
+        </Row>
     </YieldFlowComplete>
 );
