+++ suite-common/earn-stablecoin-api/src/index.ts
@@ -5,7 +5,6 @@ export * from './hooks/useAllYieldOpportunities';
 export * from './hooks/useGetYieldOpportunities';
 export * from './hooks/useEnterYieldOpportunity';
 export * from './hooks/useExitYieldOpportunity';
-export * from './hooks/useSubmitTxHash';
 export * from './hooks/useGetMerkleRewards';
 export * from './constants/vaults';
 export * from './utils/sortRewardsByUnderlyingToken';
