+++ suite-common/react-query/src/constants/mutationKeys.ts
@@ -5,7 +5,6 @@ export const commonMutationKeys = {} as const satisfies Record<string, AllowedMu
 export const desktopMutationKeys = {
     getYieldOpportunities: ['get-yield-opportunities'],
     enterYieldOpportunity: ['enter-yield-opportunity'],
-    submitTxHash: ['submit-tx-hash'],
     exitYieldOpportunity: ['exit-yield-opportunity'],
 } as const satisfies Record<string, AllowedMutationKey>;
